Conflicts:
	index.php

+++ hook.php
@@ -115,14 +115,13 @@ function cron_plugin_fusioninventory() {
 function plugin_fusioninventory_install() {
    global $DB, $LANG, $CFG_GLPI;
 
-   $version = "2.4.0";
    include (GLPI_ROOT . "/plugins/fusioninventory/install/update.php");
-   $version_detected = pluginFusioninventoryGetCurrentVersion($version);
-   if ((isset($version_detected)) AND ($version_detected != $version)) {
+   $version_detected = pluginFusioninventoryGetCurrentVersion(PLUGIN_FUSIONINVENTORY_VERSION);
+   if ((isset($version_detected)) AND ($version_detected != PLUGIN_FUSIONINVENTORY_VERSION)) {
       pluginFusioninventoryUpdate($version_detected);
    } else {
       include (GLPI_ROOT . "/plugins/fusioninventory/install/install.php");
-      pluginFusioninventoryInstall($version);
+      pluginFusioninventoryInstall(PLUGIN_FUSIONINVENTORY_VERSION);
    }
 
    return true;
@@ -810,14 +809,15 @@ function plugin_item_purge_fusioninventory($parm) {
          }
          $NetworkPort->getFromDB($parm->getField('networkports_id_2'));
          if ($NetworkPort->fields['itemtype'] == 'PluginFusioninventoryUnknownDevice') {
-            $PluginFusioninventoryUnknownDevice->getFromDB($NetworkPort->fields['items_id']);
-            if ($PluginFusioninventoryUnknownDevice->fields['hub'] == '1') {
-               $a_vlans = $NetworkPort_Vlan->getVlansForNetworkPort($NetworkPort->fields['id']);
-               foreach ($a_vlans as $vlan_id) {                  
-                  $NetworkPort_Vlan->unassignVlan($NetworkPort->fields['id'], $vlan_id);
+            if ($PluginFusioninventoryUnknownDevice->getFromDB($NetworkPort->fields['items_id'])) {
+               if ($PluginFusioninventoryUnknownDevice->fields['hub'] == '1') {
+                  $a_vlans = $NetworkPort_Vlan->getVlansForNetworkPort($NetworkPort->fields['id']);
+                  foreach ($a_vlans as $vlan_id) {
+                     $NetworkPort_Vlan->unassignVlan($NetworkPort->fields['id'], $vlan_id);
+                  }
+                  $a_hubs[$NetworkPort->fields['items_id']] = 1;
+                  $NetworkPort->delete($NetworkPort->fields);
                }
-               $a_hubs[$NetworkPort->fields['items_id']] = 1;
-               $NetworkPort->delete($NetworkPort->fields);
             }
          }
          // If hub have no port, delete it
